+++ snowplow-demo-new/build.gradle
@@ -61,7 +61,6 @@ dependencies {
     implementation 'androidx.preference:preference:1.1.1'
     implementation 'androidx.browser:browser:1.3.0'
     
-    implementation 'com.squareup.okhttp3:okhttp:4.10.0'
     implementation "com.squareup.retrofit2:retrofit:2.9.0"
     implementation "com.squareup.retrofit2:converter-gson:2.0.0"
 }
