+++ snowplow-demo-new/src/main/java/com/snowplowanalytics/snowplow_demo_new/ui/SchemaDetailViewModel.kt
@@ -1,8 +1,9 @@
-package com.snowplowanalytics.snowplow_demo_new
+package com.snowplowanalytics.snowplow_demo_new.ui
 
 import androidx.compose.runtime.*
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
+import com.snowplowanalytics.snowplow_demo_new.IgluAPIService
 import com.snowplowanalytics.snowplow_demo_new.data.SchemaUrlParts
 import kotlinx.coroutines.launch
 import org.json.JSONObject
